Adds support for pick up and drop offs for quest system
Pick up and drop off support has been added to the quest system. There
are pick up and drop off child gameobjects under the quest manager game
object that contains all pick up and drop off locations. Thos locations
need to be added as references in the appropriate lists in the quest
manager. The pick up location is assigned based on which one is the
closest to the player. The drop off location is assigned randomly.