Adds pickup and drop off times, renames Quest.cs class to DropOff.cs
Adds pick up and drop off times to QuestManager.cs using OnTriggerStay in the PickUp.cs and DropOff.cs classes. The player will have to stay inside of the pickup/drop off locations for the amount of time specified for the pickup/drop off to be successful.
Renames Quest.cs class to DropOff.cs due to confusion (sorry lol)