Adds boss dialogue system, bug fixes
Adds the boss dialogue system. After a designer defined amount of completed normal quests, the boss can ask the player if they want to do a quest for a greater XP award. The player has to pick up and make the delivery in the specified time for the current player level.
Fixes a bug where no current quest would be set if no drop offs were in range of the current player.