Adds pause menu, menu controller support, flight speed interpolation
Adds a pause menu to the player UI that pauses the players game when
active. The player can resume the game where they left off, or they can
exit to the main menu.
Controller(gamepad) support has been added throughout all menus for navigating
and selecting buttons.
Flight speed is now interpolated based on a new property in the player
controller called 'speedChangeMultiplier' that lerps the current speed
to the wanted speed for a smooth accel/deccel.